Provide `sudo` run method helper and support with `FakeCap`
This provides a `sudo` run method (like regular `run`) so it
doesn't have to be called on the cap instance directly. This also
supports the `sudo` method with `FakeCap`. It will store it like
`run` does, but will prepend "sudo". This matches with how you
would normally write the command and allows finding them in the
`cmds_run` collection.

Closes #4