Hotfix: Update to the latest scmd
This updates to the latest scmd. This ensures cap-util works with
it and allows other repos to update to the latest scmd that are
also using cap-util.

This also locks down highline and net-ssh in the gemfile to avoid
bundler trying to pull in versions that require a newer ruby.

/cc @kellyredding