remove rake task util; replace with more generic "run cmd" util
All the rake task util was doing was making it easy to choose
which cap dir to run the cmd from and add env vars.  I reworked it
so that any generic system cmd string can be passed in - otherwise
it works just as before.

I'm doing this b/c it is a little bold to assume rake tasks will be
involved in cap deploys.  Plus this makes for a more generally useful
util as rake tasks are just a very specific kind of system cmd.