rework `RunCmd` as `BundlerCmd`
When trying to work in the run cmd in a production app the modeling
didn't feel right - it felt forced.  The run cmd proved to be too
generic to have much meaning but not generic enough to have widespread
utility (something that generic isn't really necessary anyway).

In our production apps, I found myself wanting to only use the run
cmd on cap tasks that executed system cmds that required `bundle exec`.
This prompted me to rework the run cmd as a bundler cmd.  This new
version is identical to the run cmd except it adds `bundle exec` for
you and should only be used when the cmd root contains a Gemfile and
the cmd needs a bundled load path etc.

This also adds a test helper for asserting a util is a correct bundler
cmd util.